scrapy_crawler/requirements.txt
scrapy>=2.11.0
scrapy-playwright>=0.0.34
readability-lxml>=0.8.1
playwright>=1.40.0
